# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/stats/portfolio/index": {
            "get": {
                "tags": [
                    "Portfolio Stats"
                ],
                "summary": "Get statistics for all user portfolios",
                "description": "Returns overall profit statistics for all user portfolios, including total user profit and individual portfolio stats.",
                "operationId": "5beb2c68154d6a03ca8cdaa2a9c13ea6",
                "responses": {
                    "200": {
                        "description": "User and portfolios statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userAccountStats": {
                                            "properties": {
                                                "totalProfitPrice": {
                                                    "description": "Total profit or loss in USD",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 125.35
                                                },
                                                "totalProfitPercent": {
                                                    "description": "Total profit or loss in percentage",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 14.27
                                                },
                                                "profitSide": {
                                                    "description": "Profit side: '+' for gain, '-' for loss",
                                                    "type": "string",
                                                    "enum": [
                                                        "+",
                                                        "-"
                                                    ],
                                                    "example": "+"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "porftoliosStats": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "balanceStartValue": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1000
                                                    },
                                                    "balanceCurrentValue": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 1125.35
                                                    },
                                                    "profitValuePercent": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 12.53
                                                    },
                                                    "profitValuePrice": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 125.35
                                                    },
                                                    "profitSide": {
                                                        "type": "string",
                                                        "enum": [
                                                            "+",
                                                            "-"
                                                        ],
                                                        "example": "+"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
        "/api/stats/portfolio/get/{id}": {
            "get": {
                "tags": [
                    "Portfolio Stats"
                ],
                "summary": "Get statistics for single user portfolio",
                "description": "Returns profit statistics for single user portfolios, including portfolio profit and individual transactions stats.",
                "operationId": "bb65fb0ebab839017ba2fee91ef12f44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Portfolio ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Portfolio statistics with transactions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "totalStats": {
                                            "properties": {
                                                "balanceStartValue": {
                                                    "description": "Initial portfolio balance value",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 391890.39
                                                },
                                                "balanceCurrentValue": {
                                                    "description": "Current portfolio balance value",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 509325.45
                                                },
                                                "profitValuePercent": {
                                                    "description": "Profit percentage",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 29.96
                                                },
                                                "profitValuePrice": {
                                                    "description": "Profit value in currency",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 117435.05
                                                },
                                                "profitSide": {
                                                    "description": "Profit direction: '+' for gain, '-' for loss",
                                                    "type": "string",
                                                    "enum": [
                                                        "+",
                                                        "-"
                                                    ],
                                                    "example": "+"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "transactionStats": {
                                            "description": "Statistics for individual transactions, keyed by transaction ID",
                                            "type": "object",
                                            "additionalProperties": {
                                                "properties": {
                                                    "profitValuePercent": {
                                                        "description": "Transaction profit percentage",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 14.59
                                                    },
                                                    "profitValuePrice": {
                                                        "description": "Transaction profit value in currency",
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 14.83
                                                    },
                                                    "profitSide": {
                                                        "description": "Transaction profit direction: '+' for gain, '-' for loss",
                                                        "type": "string",
                                                        "enum": [
                                                            "+",
                                                            "-"
                                                        ],
                                                        "example": "+"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "SanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "SanctumAuth": {
                "type": "apiKey",
                "description": "Введите токен в формате 'Bearer {your-token}'",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}